This is going to be a longer message.
I spent a while going down the rabbit hole of how to represent the
system objects and schema in the simplest, friendliest way possible.

This then sent me down a spiral on how keys are laid out, how system
tables get setup, how users interact with this information, and much
much more.

One of the things I struggled with was how to setup the system tables,
specifically what primary keys they used.

I also struggled with whether the columns would be nested inside the
payload of the table or as its own system table.

Then I spent some time thinking about how sequences would be used to
generate primary key values.

I also skirted around the question of how keys are to be presented.

For a table

Table {
 ID: 0
 Name: Users
}

Would you prefer a key:

/__table/Users

or one

/__table/0?

The keys for the rows on the table won't be too frightening, eg:

/Users/1

For columns we would be looking at something to the effect of:

/__columns/1

Where we wouldn't necessarily need to autogenerate its values. Would we?

I suppose an integer key would be necessary, but then I need to rework
all my tests.